# docs/swagger.yaml
basePath: /api/v1
definitions:
  auth.EmailRequest:
    properties:
      email:
        type: string
    required:
    - email
    type: object
  auth.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  auth.RegisterRequest:
    properties:
      email:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      password:
        minLength: 6
        type: string
    required:
    - email
    - first_name
    - last_name
    - password
    type: object
  auth.ResetPasswordRequest:
    properties:
      code:
        type: string
      confirm_password:
        type: string
      email:
        type: string
      new_password:
        minLength: 6
        type: string
    required:
    - code
    - confirm_password
    - email
    - new_password
    type: object
...
  parking_lot.CreateParkingLotRequest:
    properties:
      location:
        type: string
      name:
        type: string
    required:
    - location
    - name
    type: object
  parking_lot.UpdateParkingLotRequest:
    properties:
      location:
        type: string
      name:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Backend API cho hệ thống smart parking
  title: Smart Parking Backend API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Đăng nhập bằng email và mật khẩu
      parameters:
      - description: Thông tin đăng nhập
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: Đăng nhập
      tags:
      - auth
  /auth/logout:
    post:
      description: Đăng xuất tài khoản hiện tại và xóa refresh token cookie
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Đăng xuất
      tags:
      - auth
  /auth/refresh-token:
    post:
      description: Cấp lại access token từ refresh token trong cookie
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      summary: Làm mới access token
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Tạo tài khoản mới và gửi email xác thực
      parameters:
      - description: Thông tin đăng ký
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: Đăng ký tài khoản
      tags:
      - auth
  /auth/resend:
    post:
      consumes:
      - application/json
      description: Gửi lại email xác thực tài khoản
      parameters:
      - description: Email người dùng
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.EmailRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: Gửi lại email xác thực
      tags:
      - auth
  /auth/reset-password:
    patch:
      consumes:
      - application/json
      description: Đặt lại mật khẩu bằng mã xác thực hoặc token reset
      parameters:
      - description: Thông tin đặt lại mật khẩu
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.ResetPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: Đặt lại mật khẩu
      tags:
      - auth
  /auth/send-reset-password:
    post:
      consumes:
      - application/json
      description: Gửi email chứa hướng dẫn hoặc mã để đặt lại mật khẩu
      parameters:
      - description: Email người dùng
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.EmailRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: Gửi email đặt lại mật khẩu
      tags:
      - auth
  /auth/verify:
    get:
      description: Xác thực tài khoản bằng email và mã code từ link gửi qua email
      parameters:
      - description: Email
        in: query
        name: email
        required: true
        type: string
      - description: Mã xác thực
        in: query
        name: code
        required: true
        type: string
      produces:
      - text/html
      responses:
        "200":
          description: HTML verified page
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: Xác thực email
      tags:
      - auth
...
  /parking-lots:
    get:
      description: Trả về toàn bộ danh sách bãi đỗ xe
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Lấy danh sách bãi đỗ xe
      tags:
      - parking_lot
    post:
      consumes:
      - application/json
      description: Tạo mới một bãi đỗ xe trong hệ thống
      parameters:
      - description: Thông tin bãi đỗ xe
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/parking_lot.CreateParkingLotRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Tạo bãi đỗ xe
      tags:
      - parking_lot
  /parking-lots/{id}:
    get:
      description: Lấy thông tin chi tiết của một bãi đỗ xe theo ID
      parameters:
      - description: ID bãi đỗ xe
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Lấy chi tiết bãi đỗ xe
      tags:
      - parking_lot
    patch:
      consumes:
      - application/json
      description: Cập nhật thông tin bãi đỗ xe theo ID
      parameters:
      - description: ID bãi đỗ xe
        in: path
        name: id
        required: true
        type: integer
      - description: Thông tin cập nhật bãi đỗ xe
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/parking_lot.UpdateParkingLotRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Cập nhật bãi đỗ xe
      tags:
      - parking_lot
  /parking-lots/{id}/gates:
    get:
      parameters:
      - description: ID bãi đỗ xe
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Lấy danh sách cổng theo bãi đỗ
      tags:
      - parking_lot
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"